Protect against -Wundef
-Wundef detects cases where an undefined symbol is evaluated in the
context of #if and defaults to 0. While this default behavior could
sometimes be a nice shortcut it can also hide issues like lack of proper
configuration or a misspelled macro.

Furthermore, clients adding the library by CMake add_subdirectory might
be hit by the problem of -Wundef triggered in the library itself.
Protecting against it also simplifies client scenarios by not requiring
the client to disable the flag in the library builds.

Change-Id: If51cc6033a24f58d34cfdab5b8d4ede306547d08